requirements.txt
numpy>=1.23
pandas>=1.5
torch>=2.0
scikit-learn>=1.2
matplotlib>=3.7
streamlit>=1.28
PyYAML>=6.0
scipy>=1.10
h5py>=3.8     
seaborn>=0.12 
pytest>=7.0   
pytest-cov>=4.0 
